Final Commit
Ready for review. We created multiple if else statements that will allow for the current day of the month be displayed and also one that will highlight the current day of the month. An if statement was created that will allow for the schedule for the day to be displayed.